Create a separate widget for trajectory output

+++ MDANSE_GUI/Src/MDANSE_GUI/PyQtGUI/InputWidgets/FloatWidget.py
@@ -68,7 +68,7 @@ class FloatWidget(WidgetBase):
 
     @Slot(str)
     def newText(self, text: str):
-        self._value = int(text)
+        self._value = float(text)
         self.updateValue()
 
     @Slot(float)
